feat(Sampler): Implemented the min-max finder
So we are trying to find the min-max of each thing so that we can
scale down the values for the supervisor and then we can supervise
the sequences and proceed.

The next step for the SupervisedSampler is just implementing everything.

For the SeriesGAN it wouldn't be using these values in the first place,
as the (min-max) will be whatever the data was trained on and will need to be provided.