feat(SupervisedNormal): Wrote the preprocessing logic for Supervisor Pass
For the supervisor pass, there's a need to create targets from the input sequence
so that they can be passed to the model which will be able to supervise them and
create a new sequence.

So right now, I have wrote the min-max logic for the input sequence so that
it roughly matches the format the supervisor saw in training.
I then split the sequence into targets (we on purpose generate days_to_tale + lookback) to account
for the fact that some data will be lost.

The padding and stacking is also done, so now assuming everything works fine:
As I am commiting this, it builds and the tensor shape and appearance are as expected

It will be a matter of taking the input from the supervisor, stacking that and calling it a day.
Once that is implemented, it's writing the evolution logic that will come since we can now
generate sequences that hopefully capture both the mean/variance of real data as well as
some of the temporal qualities thanks to the supervisor.

If I can fix the generator, then it will be playing that role but in the meantime, this will hopefully
work just fine.

I am happy, life is good. :)
~ A Chill Guy :)