feat(evolve_portfolio): Non-Dominated Sortting Implemented
So I used the code that generates the Pareto front from the dataset.
To create repeatedly Pareto fronts from the list.

Now I need to implement the code to edit the crowding distances and
get everything in order.

The next step will be the breeding code and once that's done I'll modularize
it because it's starting to look like a bit of a mess.

I think a good plan would be to move the code that I plan to expose in its
own modules.