feat(evolve_portfolio): Implemented Crowding Distance
So now the crowding distance is implemented.
The logic checks out, but I have been bulk writing stuff
(been pretty much writing code since 9 this morning.)

Now that the crowding distance logic is implemented the rest
is to select the portfolios that perform best by
going through Pareto fronts and then crowding distances.

Now that we have this done, the next step is selecting these
portfolios until we reached the right level of elitism (exploitation)
and then fill the rest with the offspring (exploration).

I need to add a variable that allows user to set the right level
of elitism they want.