feat(evolve_portfolios): Core Logic is Written
Yahoo after a coding marathon of something like 12 hours (I took short breaks.)
The core logic is implemented.
To summarize:
1. Sample the required number of portfolios
2. Per generation, sample a K scenarios and run each portfolios on all K-scenarios
to have a better idea of their performance.
3. Take the average of all the metrics we care about for this implementation
that is returns, volatiltiy and Sharpe ratio to have an idea of expected performance.
4. Build the Pareto fronts of non-dominated solutions
5. Going from best Pareto fronts back, fill the population for next generation
to elite_population_size (computed based on the full population_size and a
provided percentage of elites/exploitation of past generation).
6. Fill in the rest using offspring generation from these non-dominated (and less dominated)
solutions.
7. Enjoy!!!?!

At this point, I think the core logic is done.
Few refinements could be added but it should be suffice.

I just need to add the metric vectors to use for reporting
and boom, the algorithm will be complete.
After that I'll modularize it or move sections around to
make the whole thing clearer, and then go from there.

I should probably write the documentation at some point, but not making any promises
for that lol. The code should hopefully be self-documenting.

We will see if me of tomorrow agrees.
GGs me, you can now rest on a grateful universe.
(False I'll add the metric stuff before going to bed) ;P