feat, fix: Just cleaning up and refining
So caught a few bugs related to the sharpe ratio computation and the fact I was annualizing a variable
which is already annualized (risk-free rate) so I fixed that/

I also changed the days_to_sample to period_to_sample to better match with the
new time-horizon logic.

At this point besidest testing and running the simulation, the next step would be to simply run this and get results.
As time allows develop more samplers and stuff andmodularize as well.

But for today this is more than enough.