feat(main): MEANS and COVARIANCE based on data
Well I am sorta mad because I just realized that the supervising
trick while functional for the application becomes uncharted territory.

In the sense that while the paper has trained the model on financial data
and that the generator was able to train those properties well,
there is no way a simple multivariate normal will be able to capture those
non-stationary properties.

Hopefully the supervisor will be able to capture some of it.
But as is, the usability of the results outside of the analysis are
questionable.

I think my best bet would be one of two things.
The first and most simple one would be to retrain SeriesGAN
on the returns of the model rather than the actual stock prices themselves.

This would save me a step and would potentially represent less
for the model to learn as the non-stationarity of financial data
is already abstracted away.

But if not that, the second best thing is simply to not GAF and
go with this, and then validate the result and performance
of the model in isolation later.

While potentially not ideal, it might be possible that the supervisor
which also was trained on non-stationary data is powerful enough
to correct the normal. I wouldn't bet on it, but hopefully it can do it.

A third option which I just thought of would be to use the God awful generator
still for rigor. The results would still be unusable, but at the very
least once the generator is replaced we would gain a boost in performance for free.

All things considered, at this point.
It is likely better to just keep running with the supervised normal approach
which we have and then use that.

Hopefully we find interesting portfolios nonetheless.