feat(main): Added a little save thingey
It would be silly to run all this computation and to not
have anything to say for it cause we printed 5.
So I modified the code so that it just always saves
it somewhere.

I will make it take a path in the future, but for now it's good enough.