feat(Sampler::Normal): Implemented the normal logic based on log returns
So the previous code made the naive assumption that a normal distribution
could represent the flow of a normal well enough that simply supervising it
would be sufficient.

It wasn't and gave ridiculously high Sharpe ratios of like 5-6-7 and above.

This gives more reasonable results.

The next step is to refine the evolution logic.
- Implement tournament selection or a better form of
parent selection since currently we use only the elite
- Second add some hypervolume indicator to measure and track performance
of Pareto front
- Third modularize the code, I want the code to be more easily readable
cause damn.

But I am happy, genuinely so.