feat(bin->lib): Converted my main.rs to a lib.rs
Now, I am now converting the thing to become a python endpoint.
This at term will be a library that I can use in other files so
it makes sense to convert everything to a lib.rs file and work
from there.

Also made the changes to the Cargo.toml so that it's ready for PyO3 and Maturin.