feat(factor model): Worked more on the factor model implementation. There's like two functions left, but after that I should be good to go.
It was needed since the proximal learning that I will add for the evolution algorithm only really makes sense if I have high dimensional problem (lots of portfolios) and I couldn't be bothered to make a realistic dataset of 100 or so portfolios to get to a high dimensional dataset.

This is a feature to have as well as it will help make the analysis of model performance (irrespective of underlying assets) more realistic as they can be considered in abstraction.